fix: 평가 작성 시 content null이었던 문제 해결
- 평가 작성 시 내용  값이 안들어가고 있었음, db 단에 constraint 추가해서 알게 됨
- 내용 들어가도록 수정, 테스트코드에 반영

+++ src/test/java/com/guenbon/siso/support/fixture/rating/RatingWriteDTOFixture.java
@@ -10,7 +10,7 @@ public class RatingWriteDTOFixture {
 
     private String congressmanId = "congressmanId";
     private String content = "content";
-    private Float rating = 3.0f;
+    private Double rating = 3.0;
 
     public static RatingWriteDTOFixture builder() {
         return new RatingWriteDTOFixture();
@@ -26,7 +26,7 @@ public class RatingWriteDTOFixture {
         return this;
     }
 
-    public RatingWriteDTOFixture setRating(Float rating) {
+    public RatingWriteDTOFixture setRating(Double rating) {
         this.rating = rating;
         return this;
     }
